Move find_first_iso_image to the Payloads module
The function is called only from the module.
