+++ pyanaconda/modules/payloads/source/utils.py
@@ -15,17 +15,23 @@
 # License and may only be used or replicated with the express permission of
 # Red Hat, Inc.
 #
+import os
 import os.path
+import stat
 
+import blivet.util
 from blivet.arch import get_arch
 from blivet.util import mount
+from productmd import DiscInfo
 
+from pyanaconda import isys
 from pyanaconda.core.constants import SOURCES_DIR
 from pyanaconda.core.storage import device_matches
 from pyanaconda.core.path import join_paths
-from pyanaconda.payload.image import find_first_iso_image
-
 from pyanaconda.anaconda_loggers import get_module_logger
+from pyanaconda.modules.payloads.payload.dnf.tree_info import TreeInfoMetadata, \
+    TreeInfoMetadataError
+
 log = get_module_logger(__name__)
 
 
@@ -114,7 +120,7 @@ def find_and_mount_iso_image(source_path, mount_path):
     :param str mount_path: where to mount the ISO image
     :return: name of the ISO image file or empty string if ISO can't be used
     """
-    iso_name = find_first_iso_image(source_path)
+    iso_name = _find_first_iso_image(source_path)
 
     if iso_name:
         path_to_iso = _create_iso_path(source_path, iso_name)
@@ -123,6 +129,103 @@ def find_and_mount_iso_image(source_path, mount_path):
             return iso_name
 
     return ""
+
+
+def _find_first_iso_image(path, mount_path="/mnt/install/cdimage"):
+    """Find the first iso image in path.
+
+    :param str path: path to the directory with iso image(s); this also supports pointing to
+        a specific .iso image
+    :param str mount_path: path for mounting the ISO when checking it is valid
+
+    FIXME once payloads are modularized:
+      - mount_path should lose the legacy default
+
+    :return: basename of the image - file name without path
+    :rtype: str or None
+    """
+    try:
+        os.stat(path)
+    except OSError:
+        return None
+
+    arch = get_arch()
+    discinfo_path = os.path.join(mount_path, ".discinfo")
+
+    if os.path.isfile(path) and path.endswith(".iso"):
+        files = [os.path.basename(path)]
+        path = os.path.dirname(path)
+    else:
+        files = os.listdir(path)
+
+    for fn in files:
+        what = os.path.join(path, fn)
+        log.debug("Checking %s", what)
+        if not isys.isIsoImage(what):
+            continue
+
+        log.debug("Mounting %s on %s", what, mount_path)
+        try:
+            blivet.util.mount(what, mount_path, fstype="iso9660", options="ro")
+        except OSError:
+            continue
+
+        if not os.access(discinfo_path, os.R_OK):
+            blivet.util.umount(mount_path)
+            continue
+
+        log.debug("Reading .discinfo")
+        disc_info = DiscInfo()
+
+        # TODO replace next 2 blocks with:
+        #   pyanaconda.modules.payloads.source.utils.is_valid_install_disk
+        try:
+            disc_info.load(discinfo_path)
+            disc_arch = disc_info.arch
+        except Exception as ex:  # pylint: disable=broad-except
+            log.warning(".discinfo file can't be loaded: %s", ex)
+            continue
+
+        log.debug("discArch = %s", disc_arch)
+        if disc_arch != arch:
+            log.warning("Architectures mismatch in find_first_iso_image: %s != %s",
+                        disc_arch, arch)
+            blivet.util.umount(mount_path)
+            continue
+
+        # If there's no repodata, there's no point in trying to
+        # install from it.
+        if not _check_repodata(mount_path):
+            log.warning("%s doesn't have a valid repodata, skipping", what)
+            blivet.util.umount(mount_path)
+            continue
+
+        # warn user if images appears to be wrong size
+        if os.stat(what)[stat.ST_SIZE] % 2048:
+            log.warning(
+                "The ISO image %s has a size which is not "
+                "a multiple of 2048 bytes. This may mean it "
+                "was corrupted on transfer to this computer.",
+                what
+            )
+            blivet.util.umount(mount_path)
+            continue
+
+        log.info("Found disc at %s", fn)
+        blivet.util.umount(mount_path)
+        return fn
+
+    return None
+
+
+def _check_repodata(mount_path):
+    try:
+        tree_info_metadata = TreeInfoMetadata()
+        tree_info_metadata.load_file(mount_path)
+        return tree_info_metadata.verify_image_base_repo()
+    except TreeInfoMetadataError as e:
+        log.debug("Can't read install tree metadata: %s", str(e))
+        return False
 
 
 def mount_iso_image(image_path, mount_point):
